refactor(novu): move restoreStdinForConsole out of connect/ui fixes NV-8298
The main CJS build excludes connect/ui; pipeline code must not import from
that tree. Shared terminal helper now lives at connect/restore-stdin-for-console.ts.
